// src/app/models/paneles.model.ts
// Generated by https://quicktype.io

export interface Panel {
  idPanel:             number;
  idProveedor:         number;
  customerId:          string;
  tipoComunicador:     string;
  particiones:         number;
  usuarioCreacion:     number;
  usuarioModificacion: null;
  fechaCreacion:       string;
  fechaModificacion:   null;
  idPanelMbdb:         null;
  idPersona:           null;
  latitud:             number;
  longitud:            number;
  eventoAlarma:        any[];
  personaPanel:        any[];
  proveedor:           null;
  idProveedorWeb:      number;
  idUsuarioWeb:        number;
  proveedorWeb:        null;
}